==> 1_EDA.rmd <==
---
title: "What Makes a Hit Song?"
author: "Jae, Sudhish, Hagen, Wendy"
output:
  beamer_presentation:
    slide_level: 3
    theme: Boadilla
    number_sections: yes
  ioslides_presentation: default
  slidy_presentation: default
header-includes:
- \AtBeginSection[]{\begin{frame}\tableofcontents[currentsection]\end{frame}}
- \AtBeginSubsection[]{\begin{frame}\tableofcontents[currentsubsection]\end{frame}{}}
editor_options: 
  chunk_output_type: inline
  markdown: 
    wrap: 72
urlcolor: blue
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, cache = T, fig.width=6, fig.height=3)
options(scipen = 0, digits = 3)  # controls base R output
if(!require('pacman')) {
  install.packages('pacman')
}
pacman::p_load( glmnet, leaps, car, tidyverse, pROC, caret, dplyr, ggplot2,tidyr, stringr, gridExtra, cowplot)
```

#Data Prep

``` {r}
df <- read.csv('data/group1_hit_songs_dataset_raw1.csv')
names(df)
str(df)
print(head(df))
song_df <- df

#Separate column
song_df[c("song", "artist")] <- str_split_fixed(song_df$new_id, '%:%', 2)
song_df$new_id <- NULL

song_df$artists_ids <- gsub("\\[|\\]","", as.character(song_df$artists_ids))
```

###Hitness Dist
``` {r}
hitness_dist <- song_df %>% ggplot(aes(x = hitness)) +
  geom_histogram(fill = "#E48757", bins = 30) +
  labs(title = "Distribution of the Hitness of Songs", 
       x = "Score of Hitness",
       y = "Frequency") +
  theme_bw() +
  geom_vline(xintercept = median(song_df$hitness), linetype = "dotted", 
             color = "#1E0475", size = 1) +
  geom_text(aes(x=425, label="\n\nMedian = 220", y=1350), colour="red", text=element_text(size=10))

png("plots/hitness_distribution.png")
hitness_dist
dev.off()

median(song_df$hitness)
```
The distribution of "hitness" scores based on ranking the combination of peak_position and weeks_on_chart scores. The distribution appears to be strong skewed to the right.

###Duration
``` {r}
#song duration vs. hitness
duration <- song_df %>% ggplot(aes(x = duration_ms)) +
  geom_histogram(fill = "#6094B1", bins = 30) +
  labs(title = "Distribution of the Duration of Songs", 
       x = "Duration (ms)",
       y = "Frequency") +
  theme_bw()

png("plots/duration_distribution.png")
duration
dev.off()
```
Song tends to run from 1.7e+05 ms to 2.8e+05 ms.

###Popularity Hist
``` {r}
#histogram of popularity
artist_group <- song_df %>% group_by(artists_ids)
pop_dist <- artist_group %>%
  ggplot(aes(popularity)) +
  geom_histogram(fill = "#55A3FF", bins = 10) +
  labs(title = "Distribution of the Artist Popularities",
       x = "Artist Popularity", 
       y = "Frequency") +
  theme_bw()

png("plots/artist_pop_distribution.png")
pop_dist
dev.off()

#pop_arrange = song_df %>% arrange(desc(popularity))
#artist_label = filter(song_df, popularity > 9000)
```

###Top 30 hit song
``` {r}
#Most hit song
arrange_hit <- song_df %>% arrange(desc(hitness))
top_hit <- arrange_hit[1:30, ]

hit_bar <- top_hit %>% ggplot(aes(x = forcats::fct_reorder(song, -hitness),
                                  y = hitness)) +
  geom_bar(stat = "identity", aes(fill = song)) +
  theme_bw() +
  theme(legend.position = "none",
        plot.margin = ggplot2::margin(t = 5, r = 5, b = 5, l = 0, unit = "pt"), 
        axis.text.x = element_text(angle = -40, vjust = 0, hjust = 0)) +
  labs(title = "Hitness of Top 30 Songs",
       x = "Song",
       y = "Hitness") 

png("plots/top30_hitness.png")
hit_bar
dev.off()
```

###Pop vs. Hitness
``` {r}
#pop vs. hitness
artist_pop <- song_df %>% ggplot(aes(x = popularity, y = hitness)) +
  geom_point(color = "#4B3FE3") +
  labs(title = "Hitness vs Artist Popularity", 
       x = "Artist Popularity", 
       y = "Hitness") +
  geom_smooth(method = "lm") +
  theme_bw() +
  theme(legend.position = "none")

png("plots/hitness_vs_artist_pop.png")
artist_pop
dev.off()

#With top hit song
hit_song <- top_hit$song
top3 <- arrange_hit[1:3, ]
top3 <- c(top3$song)

pop_w_top <- song_df %>% ggplot(aes(x = popularity, y = hitness)) +
  geom_point(color = "gray") + 
  geom_point(data = subset(top_hit, song %in% hit_song),
             aes(col = song),
             size = 5,
             shape = 13) +
  labs(title = "Hitness vs Artist Popularity (Top 30 Highlighted)", 
       x = "Artist Popularity", 
       y = "Hitness") +
  geom_smooth(method = "lm") +
  theme_bw() + 
  geom_text(data = subset(top_hit, song %in% top3), 
            aes(label = song),
            show.legend = FALSE, 
            position = position_dodge(width = 1),
            vjust = -1.3) +
  theme(legend.position = "none")

png("plots/hitness_vs_artist_pop_(highlighted_top_30).png")
pop_w_top 
dev.off()
```
There is a slight positive correlation between popularity of an artist and the hitness score of a song. This model predicts that for every increase in one point of artist's popularity, hitness increases by 1.51 point.

###Dance vs. Hitness
``` {r}
#dance vs hitness
dance <- song_df %>% ggplot(aes(x = danceability, y = hitness)) +
  geom_point(color = "#B27BDF") + 
  labs(title = "Hitness vs Danceability", 
       x = "Danceability", 
       y = "Hitness") +
  geom_smooth(method = "lm") +
  theme_bw() 

png("plots/hitness_vs_danceability.png")
dance
dev.off()

summary(lm(hitness ~ danceability, song_df))
``` 
There appears to be a positive correlation between danceability of a song and its hitness.

###Acousticness vs. Hitness
``` {r}
#acousticness vs hitness
acoustic <- song_df %>% ggplot(aes(x = acousticness, y = hitness)) +
  geom_point(color = "#E5CC66") + 
  labs(title = "Hitness vs Acousticness", 
       x = "Acousticness", 
       y = "Hitness") +
  geom_smooth(method = "lm") +
  theme_bw()

png("plots/hitness_vs_acousticness.png")
acoustic
dev.off()

summary(lm(hitness ~ acousticness, song_df))
```
There is a negative correlation between acousticness of a song and its hitness.

###Valence vs. Hitness
``` {r}
#valence
valence_dist <- song_df %>% ggplot(aes(x = valence, y = hitness)) +
  geom_point(color = "#EC7666") + 
  labs(title = "Hitness vs Valence", 
       x = "Valence", 
       y = "Hitness") +
  geom_smooth(method = "lm") +
  theme_bw()

png("plots/hitness_vs_valence.png")
valence_dist
dev.off()

summary(lm(hitness ~ valence, song_df))
```
There appears to be weak negative correlation between score of positivity and hitness of a song.

###Speechiness vs. Hitness
``` {r}
#speechiness
speech <- song_df %>% ggplot(aes(x = speechiness, y = hitness)) +
  geom_point(color = "#5DC08F") + 
  labs(title = "Hitness vs Speechiness", 
       x = "Speechiness", 
       y = "Hitness") +
  geom_smooth(method = "lm") +
  theme_bw()

png("plots/hitness_vs_speechiness.png")
speech
dev.off()

summary(lm(hitness ~ speechiness, song_df))
```
There is a negative correlartion between speechiness and the level of hitness of a song.

###Energy vs. Hitness
``` {r}
#energy vs. hitness
energy <- song_df %>% ggplot(aes(x = energy, y = hitness)) +
  geom_point(color = "#E799E9") + 
  labs(title = "Hitness vs Energy", 
       x = "Energy", 
       y = "Hitness") +
  geom_smooth(method = "lm") +
  theme_bw()

png("plots/hitness_vs_energy.png")
energy
dev.off()
```

###Liveness vs. Hitness
``` {r}
#liveness vs. hitness
liveness <- song_df %>% ggplot(aes(x = liveness, y = hitness)) +
  geom_point(color = "#2DB9F4") + 
  labs(title = "Hitness vs Liveness", 
       x = "Liveness", 
       y = "Hitness") +
  geom_smooth(method = "lm") +
  theme_bw()

png("plots/hitness_vs_liveness.png")
liveness
dev.off()
```

``` {r}
#hitness vs sentiment
sentiment <- song_df %>% ggplot(aes(x = sentiment, y = hitness)) +
  geom_point(color = "#2DB9F4") + 
  labs(title = "Hitness vs Sentiment", 
       x = "Sentiment", 
       y = "Hitness") +
  geom_smooth(method = "lm") +
  theme_bw()

png("plots/hitness_vs_sentiment.png")
sentiment
dev.off()
```

``` {r}
#hitness vs loudness
loudness <- song_df %>% ggplot(aes(x = loudness, y = hitness)) +
  geom_point(color = "#2DB9F4") + 
  labs(title = "Hitness vs Loudness", 
       x = "Loudness", 
       y = "Hitness") +
  geom_smooth(method = "lm") +
  theme_bw()

png("plots/hitness_vs_loudness.png")
loudness
dev.off()
```

``` {r}
#hitness vs instrumentalness
instrumentalness <- song_df %>% ggplot(aes(x = instrumentalness, y = hitness)) +
  geom_point(color = "#2DB9F4") + 
  labs(title = "Hitness vs Instrumentalness", 
       x = "Instrumentalness", 
       y = "Hitness") +
  geom_smooth(method = "lm") +
  theme_bw()

png("plots/hitness_vs_instrumentalness.png")
instrumentalness
dev.off()
```

###Mode
``` {r}
#mode
mode_vs_hit <- song_df
mode_vs_hit$mode <- as.factor(mode_vs_hit$mode)
graph <- mode_vs_hit %>% group_by(mode)
mode_gr <- ggplot(graph, aes(x = mode, y = hitness)) +
  geom_boxplot(aes(fill=mode))+
  labs(title = "Hitness vs Mode", x = "Mode", y = "Hitness") +
  theme_bw()

png("plots/hitness_vs_mode.png")
mode_gr
dev.off()
```
Although there appeas to be weak correlation between mode of a song and its hitness, mode = 1 seems to have higher median hitness and more outliers with high hitness.

###Time Signature vs. Hitness
``` {r}
#time signature vs. hitness
time_sig <- song_df
time_sig$time_signature <- as.factor(time_sig$time_signature)
year <- time_sig %>% group_by(chart_end)
sig_vs_hit <- ggplot(year, aes(x = time_signature, y = hitness)) +
  geom_boxplot(aes(fill=time_signature))+
  labs(title = "Hitness vs Time Signature", x = "Time Signature", y = "Hitness") +
  theme_bw()

png("plots/hitness_vs_time_signature.png")
sig_vs_hit
dev.off()
```
A 4 4 time signature seems to have more outliers which correlate with higher hitness.

###Key vs. hitness
``` {r}
key_gr <- song_df
key_gr$key <- as.factor(key_gr$key)
gr <- key_gr %>% group_by(key)
key <- ggplot(key_gr, aes(x = key, y = hitness)) +
  geom_bar(stat = "identity", aes(fill=key))+
  labs(title = "Hitness vs Key", x = "Key", y = "Hitness") +
  theme_bw()

png("plots/hitness_vs_key.png")
key
```

# Graph combine
``` {r}
pop <- grid.arrange(artist_pop, pop_w_top, ncol=2)
characteristics <- grid.arrange(speech, liveness, acoustic, dance, energy, valence_dist, ncol = 3, nrow = 2)
cat <- grid.arrange(sig_vs_hit, mode_gr, ncol=2)
dist <- grid.arrange(duration, pop_dist, ncol=2)
```

From regression line, we see a positive correlation between hitness of a song and popularity of artists. However, this might not be clear because of the variability of the data. Hence, we zoom in on the top 30 song with highest hitness points. Fro mthis we observe a clear positive relationship.



Overall, every characteristics of the songs seems to be influencing the hitness of a song differently. For some, as x increases, hitness level decreases. While for others, hitness level increases when x increases. Need to see which factor is significant.
